Fixed delay line issue. because we are working in blocks you always save the last sample to be written to every 512th index

+++ CAWL/CAWLDelayLine.hpp
@@ -16,7 +16,7 @@ class CAWLDelayLine
 public:
     CAWLDelayLine();
     ~CAWLDelayLine();
-    void setDelayTime(unsigned int delayTime);
+    void setDelayTime(float delayTime);
     float processNextSample(float currSample);
     
 private:
